backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-decouple>=3.8
requests>=2.31
beautifulsoup4>=4.12
gunicorn>=22.0
whitenoise>=6.7
dj-database-url>=2.1
